+++ translator/tocwconfig/sampleConfig/appsignals_and_ecs_config.yaml
@@ -115,20 +115,6 @@ exporters:
                   separator: ;
               metric_name_selectors:
                 - ^.*$
-            - dimensions:
-                - - environment
-                  - exception
-                  - operation
-                  - service_name
-              label_matchers:
-                - label_names:
-                    - Telemetry.Source
-                  regex: ^Telemend$
-                  separator: ;
-              log_group_name: /aws/telemend/metrics
-              metric_name_selectors:
-                - Count
-              namespace: Telemend
         middleware: agenthealth/logs
         namespace: ApplicationSignals
         no_verify_ssl: false
@@ -173,9 +159,9 @@ exporters:
         telemetry:
             enabled: true
             include_metadata: true
-    otlphttp/appsignals_logs:
+    otlphttp/application_signals:
         auth:
-            authenticator: headers_setter/appsignals_logs
+            authenticator: headers_setter/application_signals
         compression: gzip
         encoding: proto
         endpoint: https://logs.us-east-1.amazonaws.com
@@ -228,7 +214,7 @@ extensions:
                 mode: EC2
                 region_type: ACJ
     awscloudwatchlogsprovisioner:
-        additional_auth: sigv4auth/appsignals_logs
+        additional_auth: sigv4auth/application_signals
         logs_provision_failure_backoff: 30s
         logs_provision_timeout: 10s
         region: us-east-1
@@ -251,13 +237,13 @@ extensions:
         region: us-east-1
         role_arn: ""
         service_name: ""
-    headers_setter/appsignals_logs:
+    headers_setter/application_signals:
         additional_auth: awscloudwatchlogsprovisioner
         headers:
             - action: upsert
               from_context: aws.cloudwatch.log_group.destination
               key: x-aws-log-group
-    sigv4auth/appsignals_logs:
+    sigv4auth/application_signals:
         assume_role:
             sts_region: us-east-1
         region: us-east-1
@@ -277,7 +263,7 @@ processors:
         resolvers:
             - name: ""
               platform: ecs
-    batch/application_signals_logs:
+    batch/application_signals:
         metadata_cardinality_limit: 1000
         metadata_keys:
             - aws.cloudwatch.log_group.destination
@@ -1398,7 +1384,7 @@ processors:
                 os.version:
                     enabled: false
         timeout: 2s
-    transform/application_signals_logs:
+    transform/application_signals:
         error_mode: propagate
         flatten_data: false
         log_statements:
@@ -1441,17 +1427,17 @@ service:
         - agenthealth/traces
         - agenthealth/statuscode
         - agenthealth/logs
-        - headers_setter/appsignals_logs
-        - sigv4auth/appsignals_logs
+        - headers_setter/application_signals
+        - sigv4auth/application_signals
         - awscloudwatchlogsprovisioner
     pipelines:
-        logs/application_signals_logs:
+        logs/application_signals:
             exporters:
-                - otlphttp/appsignals_logs
+                - otlphttp/application_signals
             processors:
-                - transform/application_signals_logs
+                - transform/application_signals
                 - attributestocontext
-                - batch/application_signals_logs
+                - batch/application_signals
             receivers:
                 - otlp/grpc_0_0_0_0_4315
                 - otlp/http_0_0_0_0_4316
